+++ packages/demo-cra/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.9](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.9-alpha.2...demo-cra@1.0.9) (2019-12-02)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
+## [1.0.9-alpha.2](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.9-alpha.1...demo-cra@1.0.9-alpha.2) (2019-12-02)
+
+**Note:** Version bump only for package demo-cra
+
+
+
+
+
 ## [1.0.9-alpha.1](https://github.com/tinacms/tinacms/compare/demo-cra@1.0.9-alpha.0...demo-cra@1.0.9-alpha.1) (2019-11-28)
 
 **Note:** Version bump only for package demo-cra